renamed resampling module to data_splitting
