+++ preprocessy/pipelines/_feature_selection_pipeline.py
@@ -1,9 +1,9 @@
+from preprocessy.data_splitting import Split
 from preprocessy.encoding import Encoder
 from preprocessy.feature_selection import SelectKBest
 from preprocessy.handlenullvalues import NullValuesHandler
 from preprocessy.outliers import HandleOutlier
 from preprocessy.parse import Parser
-from preprocessy.resampling import Split
 from preprocessy.scaling import Scaler
 
 from ._base import BasePipeline
