+++ preprocessy/pipelines/_standard_pipeline.py
@@ -1,8 +1,8 @@
+from preprocessy.data_splitting import Split
 from preprocessy.encoding import Encoder
 from preprocessy.handlenullvalues import NullValuesHandler
 from preprocessy.outliers import HandleOutlier
 from preprocessy.parse import Parser
-from preprocessy.resampling import Split
 from preprocessy.scaling import Scaler
 
 from ._base import BasePipeline
